apiLogin/requirements.txt
python-jose==3.3.0
python-multipart==0.0.5
pydantic==1.9.2
fastapi==0.83.0
bcrypt==4.0.1
mongoengine==0.24.2
passlib==1.7.4
uvicorn==0.16.0
cryptography==37.0.4
# pycrypto==2.6.1